+++ tests.py
@@ -5,6 +5,7 @@ import unittest
 from tornado import testing, web
 
 from examples import SimpleRequestHandler
+from sprockets.mixins import cors
 
 
 class OptionSupportTests(testing.AsyncHTTPTestCase):
@@ -27,7 +28,9 @@ class PreflightTests(testing.AsyncHTTPTestCase):
     def get_app(self):
         return web.Application(
             [web.url('/', SimpleRequestHandler),
-             web.url('/private', SimpleRequestHandler, {'creds': True})],
+             web.url('/private', SimpleRequestHandler, {'creds': True}),
+             web.url('/reqheader', SimpleRequestHandler,
+                     {'req_headers': ['Correlation-ID']})],
             cors_origins=['http://host.example.com'],
         )
 
@@ -63,6 +66,56 @@ class PreflightTests(testing.AsyncHTTPTestCase):
                                        'Access-Control-Request-Method': 'GET'})
         self.assertEqual(response.headers['Access-Control-Allow-Credentials'],
                          'true')
+
+    def test_that_preflight_fails_for_unsupported_request_header(self):
+        response = self.fetch('/', method='OPTIONS', headers={
+            'Origin': 'http://host.example.com',
+            'Access-Control-Request-Method': 'GET',
+            'Access-Control-Request-Headers': 'Correlation-ID',
+        })
+        self.assertEqual(response.code, 403)
+
+    def test_that_preflight_succeeds_for_simple_request_headers(self):
+        request_headers = ', '.join(s.title()
+                                    for s in cors.SIMPLE_REQUEST_HEADERS)
+        response = self.fetch('/', method='OPTIONS', headers={
+            'Origin': 'http://host.example.com',
+            'Access-Control-Request-Method': 'GET',
+            'Access-Control-Request-Headers': request_headers,
+        })
+        self.assertIn(response.code, range(200, 300))
+
+    def test_that_preflight_does_not_advertise_simple_request_headers(self):
+        request_headers = ', '.join(s.title()
+                                    for s in cors.SIMPLE_REQUEST_HEADERS)
+        response = self.fetch('/', method='OPTIONS', headers={
+            'Origin': 'http://host.example.com',
+            'Access-Control-Request-Method': 'GET',
+            'Access-Control-Request-Headers': request_headers,
+        })
+        self.assertIn(response.code, range(200, 300))
+        self.assertNotIn('Access-Control-Allow-Headers', response.headers)
+
+    def test_that_preflight_does_not_advertise_simple_request_headers(self):
+        request_headers = ', '.join(s.title()
+                                    for s in cors.SIMPLE_REQUEST_HEADERS)
+        response = self.fetch('/', method='OPTIONS', headers={
+            'Origin': 'http://host.example.com',
+            'Access-Control-Request-Method': 'GET',
+            'Access-Control-Request-Headers': request_headers,
+        })
+        self.assertIn(response.code, range(200, 300))
+        self.assertNotIn('Access-Control-Allow-Headers', response.headers)
+
+    def test_that_preflight_advertises_custom_request_headers(self):
+        response = self.fetch('/reqheader', method='OPTIONS', headers={
+            'Origin': 'http://host.example.com',
+            'Access-Control-Request-Method': 'GET',
+            'Access-Control-Request-Headers': 'Correlation-ID',
+        })
+        self.assertIn(response.code, range(200, 300))
+        self.assertIn('correlation-id',
+                      response.headers['Access-Control-Allow-Headers'].lower())
 
 
 class StandardRequestTests(testing.AsyncHTTPTestCase):