+++ docs/history.rst
@@ -11,12 +11,14 @@ Version History
 
      - :mailheader:`Origin`
      - :mailheader:`Access-Control-Request-Method`
+     - :mailheader:`Access-Control-Request-Headers`
 
    - Support for the following pre-flight response headers:
 
      - :mailheader:`Access-Control-Allow-Origin`
      - :mailheader:`Access-Control-Allow-Methods`
      - :mailheader:`Access-Control-Allow-Credentials`
+     - :mailheader:`Access-Control-Allow-Headers`
 
    - Support for the following in-line response headers: